// src/app/shared/components/button/button.enum.ts
export enum ButtonUI {
  primary = 'primary',
  secondary = 'secondary',
  tertiary = 'tertiary',
  text = 'text',
  elevated = 'elevated',
  outlined = 'outlined',
  filled = 'filled',
  tonal = 'tonal',
  icon = 'icon',
}

export enum ButtonType {
  button = 'button',
  submit = 'submit',
  reset = 'reset',
}